lesson 06 - solution to challenge "incorrect"
In the challenged proposed, the "output" suggested includes multiple lines, with values... which cannot really be right. 
There is only one print statement in the code that is supposed to generate the output. 
The values indicated in the output of the challenge may be there because of previous exercices, but are not related to the problem at hand. 

Maybe the objective was to have the learner pick from one of those ? But as the solution to the challenge, it does not work. 

PS: I hope I did the "fork" stuff right (I usually do it through my own repo, this time I'm going through GitHub web interface - let's see!). If not, my apologies!